Update ClassSplit to create virtual samples for small classes
For small classes, repeat line numbers to reach to the specified line number (-minrepeat). This is useful when a class only has 1 image, but we still want to learn a concept/category for it. As the image will be randomly cropped and mirrored to generate different virtual samples, this will increase the chances that this image (and the class) is learned in DNN.