Add StatisticsImage to Tsv.exe, fix a concurrency bug in RemoveNonImage.
A concurrency bug in RemoveNonImage is fixed, which may cause number of images to be removed to the miscounted.

StatisticsImage is added to tsv.exe. The command runs as:

tsv.exe StatisticsImage -inTsv [imageTsv_file] -image [column of the image]

The command will scan through the TSV file, and gives a histogram of the size distribution of the image, and whether all images are valid.